# Use swc in Rollup builds