Uses [swc](https://swc.rs/) for Rollup builds, which greatly decreases how long it takes to bundle the app.

On my laptop I see a pretty big difference in speeds. Here's some very unscientific test results:

| Task                 | Time before | Time after | Time saved           |
| -------------------- | ----------- | ---------- | -------------------- |
| `npm run build`      | 1m52s       | 33s        | 1m19s (339% speedup) |
| `npm run compile`    | 1m27s       | 13s        | 1m14s (669% speedup) |
| `npm run fast-build` | 1m32s       | 20s        | 1m12s (460% speedup) |