@Gold856 2025-03-30
Type checking doesn't work unless you put back the TypeScript plugin. I tried doing that, and swc spit out a bunch of errors about not being able to find source maps. On top of that, my build time went up significantly to the point where it only shaved off 30 seconds off a 4 minute build.