@jonahsnider 2025-03-31
`tsc` is slow, that's why you use SWC instead. Type checking can be performed as a separate step in CI or just by using your editor to notice any compilation errors.